% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidask_changes.R
\name{bidask.changes}
\alias{bidask.changes}
\title{calculate changes in L1 bid and ask prices}
\usage{
bidask.changes(symbol, on = "minutes")
}
\arguments{
\item{symbol}{string defining the symbol to analyze}

\item{on}{timespan to use for segmenting differences, using \link[xts:endpoints]{xts::endpoints}}
}
\description{
calculate changes in L1 bid and ask prices
}
