% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.depth.R
\name{chart.depth}
\alias{chart.depth}
\title{chart market depth at a point in time for a symbol}
\usage{
chart.depth(symbol, timestamp, offset = TRUE)
}
\arguments{
\item{symbol}{string defining the symbol to analyze}

\item{timestamp}{ISO-8601 or timebased object defining the point in time you wish to chart}

\item{offset}{whether to show offers as negative, boolean, default TRUE}
}
\value{
invisible xts object showing prices and depths for timestamp
}
\description{
chart market depth at a point in time for a symbol
}
