% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_spread.R
\name{effective.spread}
\alias{effective.spread}
\title{effective bid/ask spread defined from trades}
\usage{
effective.spread(symbol, store = FALSE)
}
\arguments{
\item{symbol}{string defining the symbol to analyze}

\item{store}{boolean defining whether to store result in order book environment, default FALSE}
}
\description{
effective bid/ask spread defined from trades
}
\references{
Jong, Rindi de. 2009. \emph{The Microstructure of Financial Markets}. Cambridge. p. 93 Definition 3
}
