% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_imbalance.R
\name{lm.trade.imbalance}
\alias{lm.trade.imbalance}
\title{Trade Imbalance Model Fit}
\usage{
lm.trade.imbalance(symbol, actual.start, n.days.lookback = 5, adv.mult = 1,
  lh = "")
}
\arguments{
\item{actual.start}{start date of analysis}

\item{n.days.lookback}{number of days to look back for in sample data}

\item{adv.mult}{average daily volume multiplier for bar zize}

\item{lh}{liquid hours}

\item{x}{xts object of tick data}
}
\description{
Simple linear model fit for trade imbalance model
}
\details{
TODO
}
