% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R, R/load_trades.R
\name{load.quotes}
\alias{load.quotes}
\alias{load.trades}
\alias{load.quotes}
\title{load trades or quotes from a csv file}
\usage{
load.quotes(filename, symbol.col = "X.RIC", tz = "GMT",
  format = "\%d-\%b-\%Y \%H:\%M:\%OS", ...)

load.trades(filename, symbol.col = "X.RIC", tz = "GMT",
  format = "\%d-\%b-\%Y \%H:\%M:\%OS", ...)
}
\arguments{
\item{filename}{filename to load trades/quotes from}

\item{symbol.col}{string defining the column containing the symbol, default 'X.RIC'}

\item{tz}{timezone string as defined by \link[base:timezones]{base::timezones}, default 'GMT'}

\item{format}{character string representing the date-time specification as described in \link[base:strptime]{base::strptime}}

\item{\dots}{any passthrough parameters for \link[utils:read.csv]{utils::read.csv}}
}
\description{
load trades or quotes from a csv file
}
