% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_depth_stats.R
\name{market.depth.stats}
\alias{market.depth.stats}
\title{generate descriptive statistics and quantiles for L2 market depth}
\usage{
market.depth.stats(symbol, on)
}
\arguments{
\item{symbol}{string defining the symbol to analyze}

\item{on}{timespan to use for segmenting differences, using \link[xts:endpoints]{xts::endpoints}}
}
\description{
generate descriptive statistics and quantiles for L2 market depth
}
