% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir_utils.R
\name{newOB}
\alias{newOB}
\alias{getOB}
\alias{getOB}
\title{Create or Get Order Book Environment}
\usage{
newOB(ob = ".ob", envir = .GlobalEnv)

getOB(ob = NULL, envir = NULL)
}
\arguments{
\item{ob}{string naming the order book environment, or an environment, default '.ob'}

\item{envir}{environment to use as the parent for the order book, default .GlobalEnv}
}
\value{
pointer to the order book environment
}
\description{
Create or Get Order Book Environment
}
\examples{

.ob <- newOB()
is.environment(.ob)
# TRUE

.ob <- getOB()
is.environment(.ob)
#TRUE

}
