% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoted_spread.R, R/ticks_wide.R
\name{quoted.spread}
\alias{quoted.spread}
\alias{ticks.wide}
\title{calculate the quoted market bid/ask spread}
\usage{
quoted.spread(symbol, store = FALSE)

ticks.wide(symbol, tick_size = NULL, store = FALSE)
}
\arguments{
\item{symbol}{string defining the symbol to analyze}

\item{store}{boolean defining whether to store result in order book environment, default FALSE}

\item{tick_size}{numeric tick size to use as denominator, if NULL, look up using \link[FinancialInstrument:getInstrument]{FinancialInstrument::getInstrument}}
}
\value{
\code{xts} object containing the spread
}
\description{
For \code{quoted.spread}, the bid ask spread will be calculated as
}
\details{
\eqn{ L1.AskPrice - L1.BidPrice }

for \code{ticks.wide} the quoted spread will be additionally divided by the
\code{tick_size}.
}
