% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_size_stats.R
\name{trade.size.stats}
\alias{trade.size.stats}
\title{generate descriptive statistics and quantiles for trades}
\usage{
trade.size.stats(symbol, on)
}
\arguments{
\item{symbol}{string defining the symbol to analyze}

\item{on}{timespan to use for segmenting differences, using \link[xts:endpoints]{xts::endpoints}}
}
\description{
generate descriptive statistics and quantiles for trades
}
