% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmp.R
\name{wmp}
\alias{wmp}
\title{calculate weighted midpoint for a symbol}
\usage{
wmp(symbol, store = FALSE)
}
\arguments{
\item{symbol}{string defining the symbol to analyze}

\item{store}{boolean defining whether to store result in order book environment, default FALSE}
}
\description{
calculate weighted midpoint for a symbol
}
